Split d2x watchers from r2
d2x watchers are currently independent of the r2 watchers. Until d2x is better integrated in `Thing` and supports `registerThing`, keeping them together is too bit messy.
