+++ lib/core/init.js
@@ -7,7 +7,9 @@ import {
 	waitFor,
 	waitForChild,
 	waitForEvent,
-	newSitetable,
+	initR2Watcher,
+	initD2xWatcher,
+	isAppType,
 } from '../utils';
 import { _loadModuleOptions } from './options/options';
 import { _loadModulePrefs, _runModuleStage } from './modules/modules';
@@ -82,10 +84,13 @@ export const always: Promise<void> = loadOptions
 export const beforeLoad: Promise<void> = loadOptions
 	.then(() => _runModuleStage('beforeLoad'));
 
-export const go: Promise<*> = Promise.all([beforeLoad, bodyReady])
-	.then(() => Promise.all([newSitetable(document.body), _runModuleStage('go')]));
+export const watchers: Promise<*> = Promise.all([beforeLoad, bodyReady])
+	.then(() => isAppType('d2x') ? initD2xWatcher() : initR2Watcher());
 
-export const afterLoad: Promise<void> = Promise.all([go, loadComplete])
+export const go: Promise<*> = Promise.all([beforeLoad, bodyReady])
+	.then(() => _runModuleStage('go'));
+
+export const afterLoad: Promise<void> = Promise.all([go, watchers, loadComplete])
 	.then(() => _runModuleStage('afterLoad'));
 
 bodyStart.then(() => BodyClasses.add());
