+++ lib/utils/index.js
@@ -169,10 +169,15 @@ export {
 
 export {
 	newSitetable,
+	initR2Watcher,
 	watchForElements,
 	watchForThings,
-	watchForRedditEvents,
 } from './watchers';
+
+export {
+	initD2xWatcher,
+	watchForRedditEvents,
+} from './watchers_d2x';
 
 export * as BodyClasses from './bodyClasses';
